requirements.txt
Flask==3.0.3
mysql_connector_repackaged==0.3.1
pandas==2.0.3
